requirements.txt
pandas
newspaper3k
